feat: add solution

+++ 09 Приведение примитивных типов расширение и сужение/task1007/src/task1007/Solution.java
@@ -19,15 +19,19 @@ Requirements:
 6. Метод main() должен содержать переменную f типа float.
 7. Значения переменных менять нельзя. Можно добавлять только операторы приведения типа.
 8. Программа должна выводить текст 1234567.*/
-*/
+
 
 public class Solution {
     public static void main(String[] args) {
-        long l = (byte) 1234_564_890L;
-        int x = (byte) 0b1000_1100_1010;
-        double m = (byte) 110_987_654_6299.123_34;
-        float f = (byte) l++ + 10 + ++x - (float) m;
+        long l = 1234_564_890L;
+        int x =  0b1000_1100_1010;
+//        System.out.println(x);
+        double m = (byte)110_987_654_6299.123_34;
+//        System.out.println(m);
+        float f = (float) l++ + 10 + ++x - (float) m;
+//        System.out.println(f);
         l = (long) f / 1000;
         System.out.println(l);
+
     }
 }
